requirements.txt
google==2.0.3
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
google-search-results==1.7.1
googleapis-common-protos==1.51.0
grpcio==1.27.2
gTTS==2.1.1
gTTS-token==1.1.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
openpyxl==3.0.3
pickleshare==0.7.5
Pillow==6.2.0
pyttsx3==2.87
pytz==2017.3
SpeechRecognition==3.8.1
tensorboard==2.1.0
tensorflow==1.14.0
tensorflow-estimator==2.1.0
